// src/models/infosproapi-global.model.ts
import { InfosProSkills } from './infospro-skills.model';

export class InfosProApiGlobal {
    id: number;
    logo: string;
    name: string;
    abus: string;
    description: string;
    skills: InfosProSkills[];
    email: string;
    mobile: string;
    phone: string;
    facebook_url: string;
    linkedin_url: string;
    website_url: string;
    twitter_url: string;
    created_at: string;
    updated_at: string;
    latitude: number;
    longitude: number;
    monday_start: string;
    monday_end: string;
    tuesday_start: string;
    tuesday_end: string;
    wednesday_start: string;
    wednesday_end: string;
    thursday_start: string;
    thursday_end: string;
    friday_start: string;
    friday_end: string;
    saturday_start: string;
    saturday_end: string;
    sunday_start: string;
    sunday_end: string;

}